refactor(server): split ws to components
